+++ react/countries.js
@@ -1,56 +1,110 @@
 // countries
 import defaultRules from './country/default'
+import ARE from './country/ARE'
 import ARG from './country/ARG'
 import AUS from './country/AUS'
+import AUT from './country/AUT'
+import BEL from './country/BEL'
+import BGR from './country/BGR'
 import BOL from './country/BOL'
 import BRA from './country/BRA'
 import CAN from './country/CAN'
 import CHL from './country/CHL'
 import COL from './country/COL'
 import CRI from './country/CRI'
+import CZE from './country/CZE'
+import DEU from './country/DEU'
+import DNK from './country/DNK'
+import DOM from './country/DOM'
 import ECU from './country/ECU'
 import ESP from './country/ESP'
+import FIN from './country/FIN'
 import FRA from './country/FRA'
 import GBR from './country/GBR'
 import GRC from './country/GRC'
 import GTM from './country/GTM'
+import HRV from './country/HRV'
 import IDN from './country/IDN'
+import IND from './country/IND'
+import IRL from './country/IRL'
+import ITA from './country/ITA'
 import KOR from './country/KOR'
 import MEX from './country/MEX'
+import NIC from './country/NIC'
+import NLD from './country/NLD'
 import NZL from './country/NZL'
 import PER from './country/PER'
+import POL from './country/POL'
+import PRI from './country/PRI'
 import PRT from './country/PRT'
 import PRY from './country/PRY'
 import ROU from './country/ROU'
+import RUS from './country/RUS'
+import SGP from './country/SGP'
+import SLV from './country/SLV'
+import SMR from './country/SMR'
+import SRB from './country/SRB'
+import SVK from './country/SVK'
+import SWE from './country/SWE'
+import THA from './country/THA'
+import UKR from './country/UKR'
 import URY from './country/URY'
 import USA from './country/USA'
 import VEN from './country/VEN'
+import ZAF from './country/ZAF'
 
 export default {
   defaultRules,
+  ARE,
   ARG,
   AUS,
+  AUT,
+  BEL,
+  BGR,
   BOL,
   BRA,
   CAN,
   CHL,
   COL,
   CRI,
+  CZE,
+  DEU,
+  DNK,
+  DOM,
   ECU,
   ESP,
+  FIN,
   FRA,
   GBR,
   GRC,
   GTM,
+  HRV,
   IDN,
+  IND,
+  IRL,
+  ITA,
   KOR,
   MEX,
+  NIC,
+  NLD,
   NZL,
   PER,
+  POL,
+  PRI,
   PRT,
   PRY,
   ROU,
+  RUS,
+  SGP,
+  SLV,
+  SMR,
+  SRB,
+  SVK,
+  SWE,
+  THA,
+  UKR,
   URY,
   USA,
   VEN,
+  ZAF
 }