+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - RUS country rules.
 - IDN translations.
+- Country index list.
 
 ## [4.20.0] - 2023-05-05